+++ src/main/kotlin/dev/thynanami/idea/typst/TypstIcons.kt
@@ -4,6 +4,5 @@ import com.intellij.openapi.util.IconLoader.getIcon
 import javax.swing.Icon
 
 object TypstIcons {
-  val TYPST_FILE: Icon = getIcon("/icons/typstFiletypeIcon.svg", TypstIcons::class.java)
-  val WIDGET_ICON: Icon = getIcon("/icons/typstFiletypeIcon.svg", TypstIcons::class.java)
+  val TYPST: Icon = getIcon("/icons/typstFiletypeIcon.svg", TypstIcons::class.java)
 }