+++ src/main/kotlin/dev/thynanami/idea/typst/TypstLanguage.kt
@@ -3,3 +3,7 @@ package dev.thynanami.idea.typst
 import com.intellij.lang.Language
 
 object TypstLanguage : Language("Typst")
+
+object TypstCodeLanguage : Language("TypstCode") {
+    override fun getDisplayName() = "Typst (Code Mode)"
+}